\name{testdotC-package}
\alias{testdotC-package}
\alias{testdotC}
\docType{package}
\title{
\packageTitle{testdotC}
}
\description{
\packageDescription{testdotC}
}
\details{

This package is an attempt at learning the .C interface function. I
call a simple C routine eventually. The goal is to incorporate C
code into the rtfSweave package.

}
\author{
\packageAuthor{testdotC}

Maintainer: \packageMaintainer{testdotC}
}

\keyword{ package }

